# back/central/shared/docs/auth/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Servicio REST para la gestión de reservas multi-restaurante.",
        "title": "Restaurant Reservation API",
        "termsOfService": "https://ejemplo.com/terminos",
        "contact": {
            "name": "Equipo de Backend",
            "email": "backend@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:3050",
    "basePath": "/api/v1",
    "paths": {
        "/auth/business-token": {
            "post": {
                "security": [
                    {
                        "BusinessTokenAuth": []
                    }
                ],
                "description": "Genera un token específico para un business basado en el token principal del usuario. Para super admins (scope platform), usar business_id = 0",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Generar token de business",
                "parameters": [
                    {
                        "description": "Datos del business (business_id = 0 para super admin)",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.GenerateBusinessTokenRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Token generado exitosamente",
                        "schema": {
                            "$ref": "#/definitions/response.GenerateBusinessTokenSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Datos de entrada inválidos",
                        "schema": {
                            "$ref": "#/definitions/response.GenerateBusinessTokenErrorResponse"
                        }
                    },
                    "401": {
                        "description": "No autorizado",
                        "schema": {
                            "$ref": "#/definitions/response.GenerateBusinessTokenErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Business no encontrado o sin acceso",
                        "schema": {
                            "$ref": "#/definitions/response.GenerateBusinessTokenErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "$ref": "#/definitions/response.GenerateBusinessTokenErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/permissions/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Obtiene un permiso específico por su ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Permissions"
                ],
                "summary": "Obtener permiso por ID",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "ID del permiso",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Permiso obtenido exitosamente",
                        "schema": {
                            "$ref": "#/definitions/response.PermissionSuccessResponse"
                        }
                    },
                    "400": {
                        "description": "ID inválido",
                        "schema": {
                            "$ref": "#/definitions/response.PermissionErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Token de acceso requerido",
                        "schema": {
                            "$ref": "#/definitions/response.PermissionErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Permiso no encontrado",
                        "schema": {
                            "$ref": "#/definitions/response.PermissionErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "$ref": "#/definitions/response.PermissionErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Actualiza un permiso existente en el sistema",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Permissions"
                ],
                "summary": "Actualizar permiso",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "ID del permiso",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Datos del permiso a actualizar",
                        "name": "permission",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.UpdatePermissionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Permiso actualizado exitosamente",
                        "schema": {
                            "$ref": "#/definitions/response.PermissionMessageResponse"
                        }
                    },
                    "400": {
                        "description": "Datos de entrada inválidos",
                        "schema": {
                            "$ref": "#/definitions/response.PermissionErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Token de acceso requerido",
                        "schema": {
                            "$ref": "#/definitions/response.PermissionErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Permiso no encontrado",
                        "schema": {
                            "$ref": "#/definitions/response.PermissionErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Permiso con código duplicado",
                        "schema": {
                            "$ref": "#/definitions/response.PermissionErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "$ref": "#/definitions/response.PermissionErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Elimina permanentemente un permiso del sistema (eliminación física, no se puede recuperar)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Permissions"
                ],
                "summary": "Eliminar permiso permanentemente",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "ID del permiso",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Permiso eliminado exitosamente",
                        "schema": {
                            "$ref": "#/definitions/response.PermissionMessageResponse"
                        }
                    },
                    "400": {
                        "description": "ID inválido",
                        "schema": {
                            "$ref": "#/definitions/response.PermissionErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Token de acceso requerido",
                        "schema": {
                            "$ref": "#/definitions/response.PermissionErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Permiso no encontrado",
                        "schema": {
                            "$ref": "#/definitions/response.PermissionErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "$ref": "#/definitions/response.PermissionErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/roles/{id}/permissions": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Obtiene la lista de permisos asignados a un rol",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Roles"
                ],
                "summary": "Obtener permisos de un rol",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID del rol",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.GetRolePermissionsResponse"
                        }
                    },
                    "400": {
                        "description": "ID de rol inválido",
                        "schema": {
                            "$ref": "#/definitions/response.RoleErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Rol no encontrado",
                        "schema": {
                            "$ref": "#/definitions/response.RoleErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "$ref": "#/definitions/response.RoleErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Asigna permisos a un rol. Solo se pueden asignar permisos que pertenezcan al mismo business_type que el rol",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Roles"
                ],
                "summary": "Asignar permisos a un rol",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID del rol",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "IDs de permisos a asignar",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.AssignPermissionsToRoleRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.AssignPermissionsToRoleResponse"
                        }
                    },
                    "400": {
                        "description": "Datos de entrada inválidos",
                        "schema": {
                            "$ref": "#/definitions/response.RoleErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Rol no encontrado",
                        "schema": {
                            "$ref": "#/definitions/response.RoleErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "$ref": "#/definitions/response.RoleErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "request.AssignPermissionsToRoleRequest": {
            "type": "object",
            "required": [
                "permission_ids"
            ],
            "properties": {
                "permission_ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
...
        "request.GenerateBusinessTokenRequest": {
            "type": "object",
            "properties": {
                "business_id": {
                    "description": "0 para super admin, \u003e0 para usuarios normales",
                    "type": "integer"
                }
            }
        },
...
        "request.UpdatePermissionRequest": {
            "type": "object",
            "required": [
                "action_id",
                "name",
                "resource_id",
                "scope_id"
            ],
            "properties": {
                "action_id": {
                    "description": "ID de la action",
                    "type": "integer",
                    "example": 1
                },
                "business_type_id": {
                    "type": "integer",
                    "example": 11
                },
                "code": {
                    "type": "string",
                    "example": "users:create"
                },
                "description": {
                    "type": "string",
                    "example": "Permite crear nuevos usuarios en el sistema"
                },
                "name": {
                    "type": "string",
                    "example": "Crear usuarios"
                },
                "resource_id": {
                    "description": "ID del resource",
                    "type": "integer",
                    "example": 1
                },
                "scope_id": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
...
        "response.AssignPermissionsToRoleResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Permisos asignados exitosamente al rol"
                },
                "permission_ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "role_id": {
                    "type": "integer",
                    "example": 1
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "response.BusinessTokenResponse": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string"
                }
            }
        },
...
        "response.GenerateBusinessTokenErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "response.GenerateBusinessTokenSuccessResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/response.BusinessTokenResponse"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "response.GetRolePermissionsResponse": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer",
                    "example": 3
                },
                "message": {
                    "type": "string",
                    "example": "Permisos del rol obtenidos exitosamente"
                },
                "permissions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/services_auth_roles_internal_infra_primary_handlers_response.PermissionResponse"
                    }
                },
                "role_id": {
                    "type": "integer",
                    "example": 1
                },
                "role_name": {
                    "type": "string",
                    "example": "Administrador"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "response.PermissionErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Error al procesar la solicitud"
                }
            }
        },
...
        "response.PermissionMessageResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Permiso creado exitosamente"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "response.PermissionSuccessResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/services_auth_permissions_internal_infra_primary_handlers_response.PermissionResponse"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "response.RoleErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Error interno del servidor"
                }
            }
        },
...
        "services_auth_permissions_internal_infra_primary_handlers_response.PermissionResponse": {
            "type": "object",
            "properties": {
                "action": {
                    "type": "string",
                    "example": "create"
                },
                "action_id": {
                    "type": "integer",
                    "example": 5
                },
                "business_type_id": {
                    "type": "integer",
                    "example": 11
                },
                "business_type_name": {
                    "type": "string",
                    "example": "Propiedad Horizontal"
                },
                "code": {
                    "type": "string",
                    "example": "users:create"
                },
                "description": {
                    "type": "string",
                    "example": "Permite crear nuevos usuarios en el sistema"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "type": "string",
                    "example": "Crear usuarios"
                },
                "resource": {
                    "type": "string",
                    "example": "users"
                },
                "resource_id": {
                    "type": "integer",
                    "example": 3
                },
                "scope_code": {
                    "type": "string",
                    "example": "system"
                },
                "scope_id": {
                    "type": "integer",
                    "example": 1
                },
                "scope_name": {
                    "type": "string",
                    "example": "Sistema"
                }
            }
        },
        "services_auth_roles_internal_infra_primary_handlers_response.PermissionResponse": {
            "type": "object",
            "properties": {
                "action": {
                    "type": "string",
                    "example": "create"
                },
                "description": {
                    "type": "string",
                    "example": "Crear usuarios"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "resource": {
                    "type": "string",
                    "example": "users"
                },
                "scope_code": {
                    "type": "string",
                    "example": "system"
                },
                "scope_id": {
                    "type": "integer",
                    "example": 1
                },
                "scope_name": {
                    "type": "string",
                    "example": "Sistema"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Token de business JWT con el prefijo **Bearer** (para todos los endpoints)",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "BusinessTokenAuth": {
            "description": "Token principal JWT con el prefijo **Bearer** (solo para /auth/business-token)",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}